v1.4.6

+++ SDNode/custom_nodes/Cup-ClipBoard/__init__.py
@@ -24,10 +24,10 @@ CUR_PATH = Path(__file__).parent
 
 def rmtree(path: Path):
     # unlink symbolic link
+    if not path.exists():
+        return
     if Path(path.resolve()).as_posix() != path.as_posix():
         path.unlink()
-        return
-    if not path.exists():
         return
     if path.is_file():
         path.unlink()
@@ -48,7 +48,10 @@ def rmtree(path: Path):
             ...
 
 def register():
-    rmtree(EXT_PATH)
+    import nodes
+    if hasattr(nodes, "EXTENSION_WEB_DIRS"):
+        rmtree(EXT_PATH)
+        return
     link_func = shutil.copytree
     if os.name == "nt":
         import _winapi
@@ -69,3 +72,4 @@ def unregister():
 register()
 atexit.register(unregister)
 NODE_CLASS_MAPPINGS = {}
+WEB_DIRECTORY = "./"
